Fix export until and add shorter video json
