+++ package.json
@@ -5,6 +5,7 @@
   "main": "main.js",
   "scripts": {
     "start": "electron .",
+    "pretest": "eslint src/*.js src/lib",
     "test": "jasmine"
   },
   "repository": "https://github.com/CatalystCode/VOTT",
@@ -31,12 +32,11 @@
     "async": "^2.1.5",
     "cntk-fastercnn": "^0.1.2",
     "crypto-js": "^3.1.9-1",
-    "del": "^3.0.0",
     "electron": "^1.4.1",
     "electron-window-state": "^4.0.2",
     "remote": "^0.2.6",
-    "remove": "^0.1.5",
     "replace": "^0.3.0",
+    "rimraf": "^2.6.2",
     "send": "^0.16.2",
     "tfrecord": "^0.2.0"
   },
